Contato finalizado

+++ AmigoFielProject/src/app/components/contato/contato.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-contato',
@@ -20,10 +21,25 @@ export class ContatoComponent {
   }
 
   EnviarMsm(){
-    this.DadosContato.nome = this.NomeInput
-    this.DadosContato.email = this.EmailInput
-    this.DadosContato.numero= this.NumInput
-    this.DadosContato.mensagem = this.MsmInput
-    localStorage.setItem('ContatoUser', JSON.stringify(this.DadosContato))
+    if(this.NomeInput == "" || this.EmailInput == "" || this.NumInput == "" || this.MsmInput == ""){
+      Swal.fire({
+        title: 'Erro',
+        text: 'Preencha todo o formulário antes de enviar',
+        icon: 'error',
+        confirmButtonColor: '#003366'
+      });
+    }else{
+      this.DadosContato.nome = this.NomeInput
+      this.DadosContato.email = this.EmailInput
+      this.DadosContato.numero= this.NumInput
+      this.DadosContato.mensagem = this.MsmInput
+      localStorage.setItem('ContatoUser', JSON.stringify(this.DadosContato))
+      Swal.fire({
+        title: 'Sucesso!',
+        text: 'Mensagem enviada com sucesso. Em breve entraremos em contato.',
+        icon: 'success',
+        confirmButtonColor: '#003366'
+      });
+    }
   }
 }
